+++ src/layers/Layer.js
@@ -234,7 +234,6 @@ class Layer extends EventEmitter {
   }
 }
 
-Layer.CallbackPromise = CallbackPromise;
 
 module.exports = Layer;